auto_framework/requirements.txt
allure-pytest==2.8.6
allure-python-commons==2.8.6
altgraph==0.17
apipkg==1.5
atomicwrites==1.4.0
attrs==19.3.0
cachetools==4.2.2
certifi==2020.6.20
cffi==1.15.0
chardet==3.0.4
colorama==0.4.3
cssselect==1.1.0
cssutils==2.3.0
execnet==1.7.1
future==0.18.2
idna==2.10
importlib-metadata==1.7.0
lxml==4.6.3
more-itertools==8.4.0
MouseInfo==0.1.3
numpy==1.18.3
opencv-python==4.2.0.34
packaging==20.4
pefile==2021.5.24
Pillow==7.1.2
pluggy==0.13.1
premailer==3.10.0
py==1.9.0
PyAutoGUI==0.9.50
pycparser==2.20
pycryptodome==3.10.1
PyExecJS==1.5.1
PyGetWindow==0.0.9
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.2
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyodbc==4.0.30
pyparsing==2.4.7
pyperclip==1.8.2
PyQt5==5.15.2
PyQt5-sip==12.9.0
PyRect==0.1.4
PyScreeze==0.1.27
pytest==5.4.3
pytest-forked==1.2.0
pytest-html==2.1.1
pytest-metadata==1.10.0
pytest-repeat==0.9.1
pytest-rerunfailures==9.0
pytest-xdist==1.32.0
PyTweening==1.0.3
pywin32-ctypes==0.2.0
PyYAML==5.3.1
requests==2.25.1
requests-toolbelt==0.9.1
selenium==3.141.0
setuptools-scm==4.1.2
six==1.15.0
urllib3==1.25.10
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
yagmail==0.14.256
zipp==3.1.0
